Refactor rendering :info flash messages
