+++ apps/concierge_site/lib/templates/admin/my_account/edit.html.eex
@@ -1,5 +1,6 @@
 <h1>Account</h1>
 <%= flash_error(@conn) %>
+<%= flash_info(@conn) %>
 <div class="my-account-container">
   <%= form_for @changeset, admin_my_account_path(@conn, :update), fn f -> %>
     <div class="form-label">
